Fix: Tidy-Ups After Git Rebase
.clang-format-ignore
  Include additional new non-clang files used for Windows

bldscripts/mestest.sh
  Allow ${MESON_BUILD_DIR} to be an absolute path

include/pistache/emosandlibevdefs.h
  Use _IS_WINDOWS in preference to _WIN32

include/pistache/mailbox.h
  Better wording in two comments

include/pistache/meson.build
  Include winornix.h in list of header files

include/pistache/pist_timelog.h
  Use ps_strl.h rather than individually defining the strlxxx functions

include/pistache/winornix.h
  Additional comment

meson.build
  Windows checks included

src/client/client.cc
  Better assignment alignment in one place

src/common/eventmeth.cc
  Use Windows compatible clock constants

src/common/http.cc
  Better alignment in code

src/common/reactor.cc
  #endif was missed on rebase

src/common/transport.cc
  Better code alignment

src/server/listener.cc
  Avoid unneeded PST_STRERROR_R call

subprojects/cpp-httplib/httplib.h
  Revert accidental edit to subproject

tests/http_server_test.cc
  Remove merge "conflict" markers, correct use of recv_buffer

tests/request_size_test.cc
  Fix GNU-specific vs. Clang issue

tests/tcp_client.h
  Use PST_STRERROR

version.txt
  Most up to date version

winscripts/mesinstall.ps1
  Use elevated privileges in Windows for doing install

bldscripts/mestestdebug.sh
  Allow ${MESON_BUILD_DIR} to be an absolute path

transport.cc
  Put back in missing ifdef and endif that was causing an issue
  compiling in epoll mode in Linux

+++ include/pistache/mailbox.h
@@ -374,8 +374,8 @@ namespace Pistache
                 {
                     int efdread_res = READ_EFD(event_fd, &val);
 
-                    // Note: Without this (efdread_res == 0) check, there
-                    // can be an issue that shows up in the test
+                    // Note: Without the read-success check below, there can be
+                    // an issue that shows up in the test
                     // multiple_client_with_requests_to_multithreaded_server
                     // when calling run_http_server_test over and over again.
                     //
@@ -397,13 +397,14 @@ namespace Pistache
                     // the read succeeds.
                     if (efdread_res == 0)
                     { // success
-                        PS_LOG_DEBUG_ARGS("event_fd read, val %u",
-                                          (uint64_t)val);
+                        PS_LOG_DEBUG_ARGS("event_fd read, val %u", val);
+
                         if (!ret)
                         {
-                            // Have another try at pop, in case the push
-                            // happened after the first pop attempt but before
-                            // now
+                            // Have another try at pop, in case there was no
+                            // "pop" to do at the top of this function, but
+                            // then a push happened after the first pop attempt
+                            // but before the read above
                             PS_LOG_DEBUG("event_fd read, but pop was null");
                             ret = Queue<T>::pop();
                         }
